move func out of utils
