+++ finetune/lora.py
@@ -19,9 +19,7 @@ from lit_gpt.speed_monitor import estimate_flops, measure_flops
 from lit_gpt.tokenizer import Tokenizer
 from lit_gpt.utils import (
     check_valid_checkpoint_dir,
-    check_compatible_tokenization,
     chunked_cross_entropy,
-    get_batch,
     get_default_supported_precision,
     lazy_load,
     num_parameters,
@@ -124,7 +122,6 @@ def main(fabric: L.Fabric, data_dir: Path, checkpoint_dir: Path, out_dir: Path, 
         to_mlp=lora_mlp,
         to_head=lora_head,
     )
-
     checkpoint_path = checkpoint_dir / "lit_model.pth"
     fabric.print(f"Loading model {str(checkpoint_path)!r} with {config.__dict__}")
     with fabric.init_module(empty_init=False), quantization(quantize):
@@ -173,7 +170,7 @@ def train(
     tokenizer = Tokenizer(checkpoint_dir)
     max_seq_length, longest_seq_length, longest_seq_ix = get_max_seq_length(train_data)
 
-    check_compatible_tokenization(fabric, model, val_data, eval_iters, micro_batch_size, longest_seq_length) # sanity check
+    check_compatible_tokenization(fabric, model, val_data, eval_iters, longest_seq_length) # sanity check
     validate(fabric, model, val_data, tokenizer, longest_seq_length)  # sanity check
 
     with torch.device("meta"):
@@ -209,7 +206,7 @@ def train(
         iter_t0 = time.perf_counter()
 
         input_ids, targets = get_batch(
-            fabric, train_data, longest_seq_length, micro_batch_size, longest_seq_ix if iter_num == 0 else None
+            fabric, train_data, longest_seq_length, longest_seq_ix if iter_num == 0 else None
         )
 
         is_accumulating = (iter_num + 1) % gradient_accumulation_iters != 0
@@ -257,17 +254,13 @@ def train(
 
 @torch.no_grad()
 def validate(
-    fabric: L.Fabric, 
-    model: GPT, 
-    val_data: List[Dict], 
-    tokenizer: Tokenizer, 
-    longest_seq_length: int,
+    fabric: L.Fabric, model: GPT, val_data: List[Dict], tokenizer: Tokenizer, longest_seq_length: int
 ) -> torch.Tensor:
     fabric.print("Validating ...")
     model.eval()
     losses = torch.zeros(eval_iters)
     for k in range(eval_iters):
-        input_ids, targets = get_batch(fabric, val_data, longest_seq_length, micro_batch_size)
+        input_ids, targets = get_batch(fabric, val_data, longest_seq_length)
         logits = model(input_ids)
         loss = chunked_cross_entropy(logits[..., :-1, :], targets[..., 1:], chunk_size=0)
         losses[k] = loss.item()
@@ -292,6 +285,35 @@ def validate(
     return val_loss.item()
 
 
+def get_batch(
+    fabric: L.Fabric, data: List[Dict], longest_seq_length: int, longest_seq_ix: Optional[int] = None
+) -> Tuple[torch.Tensor, torch.Tensor]:
+    ix = torch.randint(len(data), (micro_batch_size,))
+    if longest_seq_ix is not None:
+        # force the longest sample at the beginning so potential OOMs happen right away
+        ix[0] = longest_seq_ix
+
+    input_ids = [data[i]["input_ids"].type(torch.int64) for i in ix]
+    labels = [data[i]["labels"].type(torch.int64) for i in ix]
+
+    # it's better to pad to a fixed seq length with XLA to avoid recompilation
+    max_len = max(len(s) for s in input_ids) if fabric.device.type != "xla" else longest_seq_length
+
+    def pad_right(x, pad_id):
+        # pad right based on the longest sequence
+        n = max_len - len(x)
+        return torch.cat((x, torch.full((n,), pad_id, dtype=x.dtype)))
+
+    x = torch.stack([pad_right(x, pad_id=0) for x in input_ids])
+    y = torch.stack([pad_right(x, pad_id=-1) for x in labels])
+
+    if fabric.device.type == "cuda" and x.device.type == "cpu":
+        x, y = fabric.to_device((x.pin_memory(), y.pin_memory()))
+    else:
+        x, y = fabric.to_device((x, y))
+    return x, y
+
+
 def get_max_seq_length(data: List[Dict]) -> Tuple[int, int, int]:
     # find out the minimum max_seq_length required during fine-tuning (saves memory!)
     lengths = [len(d["input_ids"]) for d in data]
@@ -303,6 +325,28 @@ def get_max_seq_length(data: List[Dict]) -> Tuple[int, int, int]:
         max_seq_length,
         longest_seq_ix,
     )
+
+
+def check_compatible_tokenization(
+        fabric: L.Fabric,
+        model,
+        val_data: List[Dict],
+        eval_iters: int,
+        longest_seq_length: int,
+    ) -> None:
+    fabric.print("Checking tokenization ...")
+
+    model.eval()
+    for _ in range(eval_iters):
+        input_ids, target_ids = get_batch(fabric, val_data, longest_seq_length)
+
+        if (input_ids > model.config.padded_vocab_size).any():
+            raise ValueError(
+                "Some input IDs exceed the vocabulary size of this model. "
+                "Please make sure you used this model's "
+                "tokenizer to prepare the dataset.")
+    model.reset_cache()
+    model.train()
 
 
 def save_lora_checkpoint(fabric, model, file_path: Path):
